CLIP-14: Пофиксил мелкие баги

+++ clipperSources/Track/Scenes/MKRScene.h
@@ -20,17 +20,17 @@
 @property(nonatomic) NSMutableArray<MKRBar *> *bars;
 @property(readonly, nonatomic) NSInteger identifier;
 
--(instancetype)initWithIdentifier:(NSInteger)identifier;
+- (instancetype)initWithIdentifier:(NSInteger)identifier;
 
--(BOOL)fillBarsWithBarManager:(MKRBarManager *)barManager;
+- (BOOL)fillBarsWithBarManager:(MKRBarManager *)barManager;
 
 /*!
   @abstract this method appends part of the composition to result and shifts resultCursorPtr
 */
--(void)makeComposition:(AVMutableComposition *)composition withBarAssets:(NSMutableDictionary *)barsAssets andWithResultCursorPtr:(CMTime *)resultCursorPtr andWithMSPQ:(NSInteger)MSPQ;
+- (void)makeComposition:(AVMutableComposition *)composition withBarAssets:(NSMutableDictionary *)barsAssets andWithResultCursorPtr:(CMTime *)resultCursorPtr andWithMSPQ:(double)MSPQ;
 
--(void)makeCompositionBar:(AVMutableComposition *)composition withBarAsset:(AVMutableComposition *)barAsset andWithBar:(MKRBar *)bar andWithResultCursorPtr:(CMTime *)resultCursorPtr andWithMSPQ:(NSInteger)MSPQ andWithBarRange:(CMTimeRange)barTimeRange usingAutoComplete:(BOOL)autoComplete;
+- (void)makeCompositionBar:(AVMutableComposition *)composition withBarAsset:(AVMutableComposition *)barAsset andWithBar:(MKRBar *)bar andWithResultCursorPtr:(CMTime *)resultCursorPtr andWithMSPQ:(double)MSPQ andWithBarRange:(CMTimeRange)barTimeRange usingAutoComplete:(BOOL)autoComplete;
 
--(void)insertTimeRange:(AVMutableComposition *)composition ofAsset:(AVAsset *)asset startAt:(CMTime)startAt duration:(CMTime)duration resultCursorPtr:(CMTime *)resultCursorPtr;
+- (void)insertTimeRange:(AVMutableComposition *)composition ofAsset:(AVAsset *)asset startAt:(CMTime)startAt duration:(CMTime)duration resultCursorPtr:(CMTime *)resultCursorPtr;
 
 @end
